requirements.txt
google-genai
dotenv
requests
aiohttp